Fix type warnings again

+++ tools/shims/msvc/sys/socket.h
@@ -14,7 +14,7 @@
 #include "../wsa_shim_perror.h" // wsa_shim_perror
 
 #ifdef _MSC_VER
-typedef long ssize_t;
+typedef long long ssize_t;
 #endif // _MSC_VER
 
 #define SHUT_RD   SD_RECEIVE
